requirements.txt
flask
flask-restplus
gunicorn
pillow
redis==2.10.6
rq==0.12.0
pyjwt
imageio
pymongo
numpy